HT_21/requirements/base.txt
Django==5.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
drf-spectacular==0.27.1
ipython==8.20.0
requests==2.31.0
pillow==10.2.0
